# perf: faster partition, sum of int and float is not int