* only the sum of all `int` is also `int`
* partition was very slow (traversed iterator twice, evaluated function twice). Added faster implementation, but even better is to skip the function call overhead entirely in the hot `reshape` function.


`kernprof -lv pytest -s test/external/external_test_speed_llama.py`


before:

```

codegen         mean runtime:  269.10ms, runs:   263.81,  431.94,  280.84,  241.09,  242.25,  252.06,  246.07,  243.95,  245.22,  243.79
methodcache     mean runtime:  233.63ms, runs:   238.96,  282.79,  214.99,  233.69,  218.97,  220.81,  218.90,  221.24,  227.14,  258.84
profile         mean runtime:  899.55ms, runs:   879.43,  890.44,  944.78,  869.44,  918.42,  871.21,  927.57,  889.79,  885.29,  919.17


Total time: 1.80852 s
File: /home/rvd/src/roelofvandijk/tinygrad/tinygrad/shape/shapetracker.py
Function: reshape at line 229

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
   229                                             @profile
   230                                             def reshape(self, new_shape: Tuple[Union[Node,int], ...]):
   231     81747     620009.8      7.6     34.3      new_ints, new_nodes = partition(new_shape, lambda s: isinstance(s, int))
   232     81747      20059.6      0.2      1.1      if new_nodes and all(isinstance(s, int) for s in self.shape):
   233                                                 # reshape from all int shape into shape with a variable, update the variable value
   234                                                 assert len(new_nodes) == 1 and isinstance(new_nodes[0], Variable), "only support adding one Variable to the int shape"
   235                                                 new_var, new_val = new_nodes[0], prod(self.shape) // prod(new_ints)
   236                                                 if new_var not in self.var_vals:
   237                                                   assert new_var.min <= new_val <= new_var.max, f"variable value {new_val} out of range [{new_var.min}, {new_var.max}]"
   238                                                   self.var_vals[new_var] = new_val
   239                                                 else: assert self.var_vals[new_var] == new_val, f"value conflicts, was {self.var_vals[new_var]}, set to {new_val}"
   240     81747      22387.7      0.3      1.2      elif not new_nodes: self.var_vals = {}
   241     78941      36236.5      0.5      2.0      if self.views[-1].shape == new_shape: return self
   242     78941     403295.9      5.1     22.3      assert all(is_sym_int(x) and x > 0 for x in new_shape), f"shape must be symbolic ints and can't contain 0 or negative numbers {new_shape}"
   243                                               # only check size for int shapes. we don't check symbolic here as long as the reshape itself can be done
   244     78941     435458.9      5.5     24.1      if all(isinstance(s, int) for s in self.shape) and all(isinstance(s, int) for s in new_shape):
   245     78941     108367.5      1.4      6.0        assert prod(self.shape) == prod(new_shape), f"can't reshape {self.shape} -> {new_shape}" # type: ignore  # mypy cannot resolve, all ints here
   246     78941     119891.3      1.5      6.6      new_view, extra = _reshape(self.views[-1], new_shape)
   247     71821      11979.2      0.2      0.7      if extra: self.views.append(new_view)
   248     71821      18631.0      0.3      1.0      else: self.views[-1] = new_view
   249     78941      12200.0      0.2      0.7      return self
```


after:
```
codegen         mean runtime:  249.53ms, runs:   290.72,  298.99,  233.41,  234.29,  231.03,  237.84,  255.91,  241.45,  235.90,  235.72
methodcache     mean runtime:  221.58ms, runs:   205.52,  280.67,  212.20,  212.96,  209.96,  234.57,  212.71,  215.45,  217.70,  214.07
profile         mean runtime:  859.32ms, runs:   842.68,  952.75,  867.48,  839.90,  843.67,  824.92,  871.83,  842.45,  854.53,  853.03

Total time: 1.09131 s
File: /home/rvd/src/roelofvandijk/tinygrad/tinygrad/shape/shapetracker.py
Function: reshape at line 229

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
   229                                             @profile
   230                                             def reshape(self, new_shape: Tuple[Union[Node,int], ...]):
   231     81747      22086.4      0.3      2.0      new_ints, new_nodes = [], []
   232    295042     185494.4      0.6     17.0      for s in new_shape: (new_ints if s.__class__ is int else new_nodes).append(s)
   233     81747      17897.4      0.2      1.6      if new_nodes and sum(self.shape).__class__ is int:
   234                                                 # reshape from all int shape into shape with a variable, update the variable value
   235                                                 assert len(new_nodes) == 1 and isinstance(new_nodes[0], Variable), "only support adding one Variable to the int shape"
   236                                                 new_var, new_val = new_nodes[0], prod(self.shape) // prod(new_ints)
   237                                                 if new_var not in self.var_vals:
   238                                                   assert new_var.min <= new_val <= new_var.max, f"variable value {new_val} out of range [{new_var.min}, {new_var.max}]"
   239                                                   self.var_vals[new_var] = new_val
   240                                                 else: assert self.var_vals[new_var] == new_val, f"value conflicts, was {self.var_vals[new_var]}, set to {new_val}"
   241     81747      25208.4      0.3      2.3      elif not new_nodes: self.var_vals = {}
   242     78941      38515.7      0.5      3.5      if self.views[-1].shape == new_shape: return self
   243     78941     415577.8      5.3     38.1      assert all(is_sym_int(x) and x > 0 for x in new_shape), f"shape must be symbolic ints and can't contain 0 or negative numbers {new_shape}"
   244                                               # only check size for int shapes. we don't check symbolic here as long as the reshape itself can be done
   245     78941     121332.8      1.5     11.1      if sum(self.shape + new_shape).__class__ is int:
   246     78941     106080.3      1.3      9.7        assert prod(self.shape) == prod(new_shape), f"can't reshape {self.shape} -> {new_shape}" # type: ignore  # mypy cannot resolve, all ints here
   247     78941     115698.7      1.5     10.6      new_view, extra = _reshape(self.views[-1], new_shape)
   248     71821      12236.2      0.2      1.1      if extra: self.views.append(new_view)
   249     71821      18514.5      0.3      1.7      else: self.views[-1] = new_view
   250     78941      12670.3      0.2      1.2      return self
   ```