@dc-dc-dc 2023-08-21
the helpers change looks good, can you make this pr just that and move the sum changes to #1606 since its the same concept