@geohot 2023-08-21
Yea the sum thing is sketch, but I'm fine with the helpers change. I wish there was a one liner for it though.