@roelofvandijk 2023-08-22
Reworked this a bit.

The logic should be identical, but it's way faster now:

Before: `27781  105.137    0.004  583.905    0.021 shapetracker.py:230(reshape)`
After: `27781   58.578    0.002  139.876    0.005 shapetracker.py:230(reshape)`


Before:
```
codegen         mean runtime:  240.39ms, runs:   245.97,  297.38,  222.51,  224.17,  223.97,  226.17,  227.69,  271.43,  234.77,  229.83
methodcache     mean runtime:  216.14ms, runs:   198.22,  258.19,  199.87,  203.08,  206.22,  206.12,  242.51,  227.72,  211.65,  207.81
profile         mean runtime:  895.53ms, runs:   848.52,  882.69,  935.33,  885.33,  865.64,  888.95,  956.84,  890.45,  896.73,  904.84
```

After:
```
codegen         mean runtime:  187.78ms, runs:   208.70,  254.92,  184.79,  181.37,  173.03,  174.03,  176.59,  175.37,  172.13,  176.82
methodcache     mean runtime:  164.05ms, runs:   148.87,  228.05,  156.17,  158.27,  156.91,  158.15,  157.74,  158.38,  157.94,  159.98
profile         mean runtime:  749.85ms, runs:   728.19,  737.71,  819.47,  768.21,  731.52,  721.25,  760.91,  736.51,  732.48,  762.29
```