@roelofvandijk 2023-08-22
Moved the partition change to https://github.com/tinygrad/tinygrad/pull/1625.